pyproject.toml
[tool.poetry]
name = "thewall"
version = "0.1.0"
description = ""
authors = ["Mykhailo Bielan <bielanm.box@gmail.com>"]
readme = "README.md"
packages = [
    {include = "thewall", from = "."},
]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
psycopg2 = "^2.9.5"
djangorestframework = "^3.14.0"
typer = "^0.7.0"
drf-yasg = "^1.21.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
